% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackable_functions.R
\name{add_comments}
\alias{add_comments}
\title{Append my_comment to my_dt 'comments' attribute}
\usage{
add_comments(my_dt = NULL, my_comment, readme_path = NULL)
}
\arguments{
\item{my_dt}{data.table input}

\item{my_comment}{Vector of character strings}

\item{readme_path}{Optional path to which the comments will be appended.}
}
\description{
Created to attach and build on commnets for data.tables. This allows them to be passed 
  along my_dt though the stacks.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
add_comments
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
