% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{add_filler}
\alias{add_filler}
\title{Adds a filler to the text.}
\usage{
add_filler(my_string, filler = "-", total_char = get_annotation_indention())
}
\arguments{
\item{my_string}{String to have filler added to it.}

\item{filler}{Character that will be repeated to create the filler}

\item{total_char}{Integer defining how long the final string will be.}
}
\value{
The created string.
}
\description{
Adds a filler to the text.
}
