% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{announce_end_of_stack}
\alias{announce_end_of_stack}
\title{Announce the ending of a stacked function}
\usage{
announce_end_of_stack(function_name)
}
\arguments{
\item{function_name}{Character string name of function}
}
\description{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
announce_end_of_stack
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
