% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{announce_total_time}
\alias{announce_total_time}
\title{Tells the total time if given a start time.}
\usage{
announce_total_time(identifier, start_time)
}
\arguments{
\item{identifier}{Character name of identifier}

\item{start_time}{Start time of}
}
\description{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
announce_total_time
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
