% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{as.bullet}
\alias{as.bullet}
\title{Formats text as a bullet.}
\usage{
as.bullet(
  string,
  level = 1,
  width = get_annotation_width(),
  indentation = "  ",
  my_bullet = "* "
)
}
\arguments{
\item{string}{The string of text that should be bulleted over.}

\item{level}{Integer indicating what level of bullet this should be.}

\item{width}{Integer indicating the number of character that should make up a line of the annotation.}

\item{indentation}{String of spaces to use for the indentation of bullets.}

\item{bullet_levels}{}
}
\value{
The bulleted string.
}
\description{
Formats text as a bullet.
}
