% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{as.footer}
\alias{as.footer}
\title{Right justifies and add a filler to the text.}
\usage{
as.footer(my_string)
}
\arguments{
\item{my_string}{String to be right justified with filler.}
}
\value{
The right justified string with filler.
}
\description{
Right justifies and add a filler to the text.
}
