% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{as.header1}
\alias{as.header1}
\title{Adds horizontal lines above and below the text.}
\usage{
as.header1(title_text)
}
\arguments{
\item{title_text}{String to use in the header.}
}
\value{
The created character vector.
}
\description{
Adds horizontal lines above and below the text.
}
