% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{as.header2}
\alias{as.header2}
\title{Adds filler to the text.}
\usage{
as.header2(title_text)
}
\arguments{
\item{title_text}{String to use for the title.}
}
\value{
The created string.
}
\description{
Adds filler to the text.
}
