% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{balanced_folds}
\alias{balanced_folds}
\title{Get indices of a fold of data that balances the distribution of values.}
\usage{
balanced_folds(y, nfolds = 3)
}
\arguments{
\item{y}{Vector to balance}

\item{nfolds}{Integer to indicate the number of folds in which to split the data}
}
\description{
This was written by Joel Parker.
}
