% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{bgv_lab_prep_count_data_for_deseq2}
\alias{bgv_lab_prep_count_data_for_deseq2}
\title{Preps counts data prior to running gene signature.}
\usage{
bgv_lab_prep_count_data_for_deseq2(my_dt, gene_cols)
}
\arguments{
\item{my_dt}{data.table input}

\item{gene_cols}{Character vector indicating the columns that should be used in the gene signature.
Uses \code{operatable_columns}.  If \code{NULL} if will take all numeric columns.}
}
\description{
\code{bgv_lab_prep_count_data_for_deseq2} is the standard Vincent Lab protocol
  for preparing gene count data prior to running in DESeq2.
\itemize{
  \item \code{\link{round_data}} Rounds data
  \item \code{\link{require_data_in_samples}} Requires genes be in 70% of samples to keep them
  \item \code{\link{drop_low_columns log2}} Drops genes that are in the lowest 10% for mean expression.
}
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
bgv_lab_prep_count_data_for_deseq2
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
\concept{differential_expression}
