% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{bgv_lab_prep_count_data_for_gene_signature}
\alias{bgv_lab_prep_count_data_for_gene_signature}
\title{Preps counts data prior to running gene signature.}
\usage{
bgv_lab_prep_count_data_for_gene_signature(my_dt, gene_cols)
}
\arguments{
\item{my_dt}{data.table input}

\item{gene_cols}{Character vector indicating the columns that should be used in the gene signature.
Uses \code{operatable_columns}.  If \code{NULL} if will take all numeric columns.}
}
\description{
\code{bgv_lab_prep_count_data_for_gene_signature} is the standard Vincent Lab protocol
  for preparing gene count data prior to running gene signatures.
\itemize{
  \item \code{\link{require_data_in_samples}} reuires the gene be expressed in 70% of the samples
  \item \code{\link{normalize_rows_by_quartile}} upper quartile normalizes the data 
  \item \code{\link{log_transform_plus}} log2 transforms (counts + 1)
}
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
bgv_lab_prep_count_data_for_gene_signature
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
\seealso{
Other gene_signature: \code{\link{calculate_gene_signatures}}
}
\concept{gene_signature}
