% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binfotron.R
\docType{package}
\name{binfotron-package}
\alias{binfotron}
\alias{binfotron-package}
\title{Binfotron Bioinformatics Analysis Tools Suite}
\description{
The purpose of this package is to perform upper level analysis of bioinformatic data.
}
\author{
\strong{Maintainer}: Dante Bortone \email{dantebortone@gmail.com}

}
\keyword{internal}
