% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{break_sentence}
\alias{break_sentence}
\title{Inserts newlines into strings after a specified character length.}
\usage{
break_sentence(string, width = get_annotation_width())
}
\arguments{
\item{string}{}

\item{width}{}
}
\value{
The same string but broken up into lines.
}
\description{
A word that will cause a string to exceed the specified width will be used to start the next line.
}
