% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fdr.R
\name{calc_fdr}
\alias{calc_fdr}
\title{Performs an FDR calculation on the fdr_on_columns.}
\usage{
calc_fdr(
  my_dt,
  fdr_by_columns = NULL,
  fdr_method = "BH",
  fdr_on_columns = "pValue",
  readme_path = NULL
)
}
\arguments{
\item{my_dt}{A data.table or data.frame that includes all columns of data needed to do the 
correction:  \code{fdr_by_columns} & \code{fdr_on_columns}}

\item{fdr_by_columns}{Optional character vector of column names to specify what 
should be group together in the fdr correction using the data.table "by" option.  
Leaving this blank will just apply the correction to all of the data values.}

\item{fdr_method}{String to tell what method to use to FDR correct. Must be one 
of the values in \code{stats::p.adjust.methods}.}

\item{fdr_on_columns}{Character vector to indicate which column(s) should be FDR 
corrected.}

\item{readme_path}{If path is provided the comments from running this function
will be appended to the path.}
}
\value{
data.table or data.frame (depending on the input) with an FDR correction 
column for each \code{fdr_on_columns}
}
\description{
The purpose of \code{calc_fdr} is to add FDR correction columns to a data.table using
the data.table in data.table out methodology
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
calc_fdr
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
