% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_IMPRES.R
\name{calc_IMPRES}
\alias{calc_IMPRES}
\title{calc_IMPRES}
\usage{
calc_IMPRES(func_GE_df, ID_list)
}
\arguments{
\item{ID_list}{List of all sample IDs}

\item{func_GE_DF}{Gene expression matrix, each column is a gene, each row is a sample}
}
\value{
Returns data table with 2 columns: sample ID & IMPRES score
}
\description{
calc_IMPRES
}
