% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{cjust}
\alias{cjust}
\title{Applies center justification to my_string.}
\usage{
cjust(my_string, filler = " ", total_char = get_annotation_width(), pad = " ")
}
\arguments{
\item{my_string}{String to be centereded}

\item{filler}{String character to be used to space the text over.}

\item{total_char}{Integer of how many characters the text should be when done.}
}
\value{
Centered text.
}
\description{
Adds filler characters to the left and right of my_string to have the text fill the 
line up total_char characters.
}
