% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_central_clone.R
\name{cluster_group_plot}
\alias{cluster_group_plot}
\title{Saves to file a plot of the clusters within a cluster group}
\usage{
cluster_group_plot(
  scores_df,
  cluster_membership,
  central_elements_by_cluster,
  cluster_group_id,
  output_path = NA,
  my_width = 8,
  my_height = 8
)
}
\arguments{
\item{scores_df}{Dataframe with elements ( i.e. features ) x principle component scores. Expects at least a PC1 and PC2 column}

\item{cluster_membership}{Integer factor with cluster membership for each element in scores_df}

\item{central_elements_by_cluster}{Character vector with element names at locations of most central elements for each cluster and "" everywhere else}

\item{cluster_group_id}{One-length character vector name for this cluster group to be used in plot title}

\item{output_path}{Character vector with fully qualified path to which plot will be saved}

\item{my_width}{One-length numeric vector with width of output plot ( in inches )}

\item{my_height}{One-length numeric vector with height of output plot ( in inches )}
}
\value{
Returns the ggplot object
}
\description{
Saves to file a plot showing the clusters within a cluster group, colored by group membership and with the central elements labeled for each cluster
}
