% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_central_clone.R
\name{cluster_list_to_df}
\alias{cluster_list_to_df}
\title{Reformat a long list of cluster members as a data.frame}
\usage{
cluster_list_to_df(cluster_members, central_elements = NA)
}
\arguments{
\item{cluster_members}{Long list of each cluster in each cluster count and the elements that belong to that cluster ( i.e. Cluster_002_001 <- Bindea_aDC, Bindea_BCells ... ).}

\item{central_elements}{Optional list of central elements by cluster id ( i.e. Cluster_002_001 <- central element of that cluster ). If omitted, central element columns will not be added to return data.frame.}
}
\value{
Returns the data.frame containing cluster membership for each group of clusters in the cluster_members list and optionally the central element of each cluster.
}
\description{
Reformats long-form list of cluster members as a data.frame with all elements as rows and one or two columns per cluster indicating cluster membership and optionally central elements for each cluster
}
