% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_central_clone.R
\name{collapse_identical_clusters}
\alias{collapse_identical_clusters}
\title{Remove identical clusters}
\usage{
collapse_identical_clusters(c_list, cluster_id_width = NA)
}
\arguments{
\item{c_list}{Named list with lists of clusters in which to look for duplicates. Assumes the cluster within each named item are all the same size and that cluster elements are all sorted the same ( i.e. all alphabetically sorted ).}

\item{cluster_id_width}{Width of the cluster id's}
}
\value{
Returns original list of clusters with duplicates stored in cluster_ids attribute of each unique cluster. Names of list for each size cluster are Cluster_{cluster-#-this_size}x{#-instances-this-cluster} ( i.e. Cluster_001x24 )
}
\description{
Preserves only one cluster instance per unique element combination
}
