% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_genes.R
\name{convert_gmt_file}
\alias{convert_gmt_file}
\title{Create gene lookup}
\usage{
convert_gmt_file(input_path, input_type, output_type, output_path,
  db_object = NULL, na_string = "NA", input_readme_path = NULL)
}
\arguments{
\item{input_path}{Path to the gmt file you would like ot convert.}

\item{input_type}{Typically ENTREZID or SYMBOL. See decription.}

\item{output_type}{Typically SYMBOL or ENTREZID. See decription.}

\item{output_path}{Full path to output gmt file.}

\item{db_object}{AnnotationDb object. If left null it will be set to  \code{org.Hs.eg.db}. Passed to 
\code{\link{create_gene_lookup}}.}

\item{na_string}{String to fill in the na positions of a gmt file. Typically "NA" or "".}

\item{input_readme_path}{Optional path to readme to import into this gmt's readme.}
}
\value{
None
}
\description{
Uses \code{\link{create_gene_lookup}} to convert one type of gmt file to another.
If using the default value for \code{db_object}(\code{org.Hs.eg.db}) the values for the inputs 
and outputs include any listed in \url{https://www.bioconductor.org/help/course-materials/2014/useR2014/Integration.html},
although the most common will likely be: ENTREZID, ACCNUM, ALIAS, UNIGENE, ENSEMBL, ENSEMBLPROT, 
ENSEMBLTRANS, GENENAME, UNIPROT, OMIM, UCSCKG, SYMBOL
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
convert_gmt_file
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
