% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackable_functions.R
\name{convert_piped_col_names_to_single_names}
\alias{convert_piped_col_names_to_single_names}
\title{Convert piped column names into one of the elements of the name.}
\usage{
convert_piped_col_names_to_single_names(
  my_dt = NULL,
  col_names = NULL,
  my_summary = NULL,
  position = 2,
  sample_key = get_default_sample_key(),
  readme_path = NULL
)
}
\arguments{
\item{my_dt}{data.table input}

\item{col_names}{Vector of character strings to name the columns that will have this 
operation performed on them.  Uses \code{\link{operatable_columns}}}

\item{my_summary}{Character string to change the default comment that will be appended 
to my_dt.}

\item{position}{Integer indicating which pipe element do you want}

\item{sample_key}{Character string to specify the column that is the sample key. This 
column will not be operated upon.}

\item{readme_path}{Optional path to which the comments will be appended.}
}
\description{
Converts piped column names into one of the elements of the name.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
convert_piped_col_names_to_single_names
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
