% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_genes.R
\name{create_gene_lookup}
\alias{create_gene_lookup}
\title{Create gene lookup}
\usage{
create_gene_lookup(
  input_values,
  input_type,
  output_type,
  sep = "\\t",
  db_object = NULL
)
}
\arguments{
\item{input_values}{Character vecter of names to lookup}

\item{input_type}{String of the input values. See description for common options.}

\item{output_type}{String of the ouput values. See description for common options.}

\item{sep}{String to indicate how gene names should be collapsed.  a value of 
\code{NULL} will return just the first item of the returned results}

\item{db_object}{AnnotationDb object.  If left null it will be set to \code{org.Hs.eg.db}.}
}
\value{
Named array that can be used to lookup gene values
}
\description{
Uses \code{\link[AnnotationDbi]{select}} to produce a lookup array for converting gene names
If using the default value for \code{db_object}(\code{org.Hs.eg.db}) the values for the inputs 
and outputs include any listed in \url{https://www.bioconductor.org/help/course-materials/2014/useR2014/Integration.html},
although the most common will likely be: ENTREZID, ACCNUM, ALIAS, UNIGENE, ENSEMBL, ENSEMBLPROT, 
ENSEMBLTRANS, GENENAME, UNIPROT, OMIM, UCSCKG, SYMBOL
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
create_gene_lookup
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
