% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_decoding.R
\name{decode_clms}
\alias{decode_clms}
\title{Decodes data.frame or data.table using decode_special_char}
\usage{
decode_clms(my_df, skip_clms = c(), replace_existing_clm = FALSE)
}
\arguments{
\item{my_df}{Input date.frame or data.table}

\item{skip_clms}{character vector of columns to skip}

\item{replace_existing_clm}{Boolean to indicate whether encoded columns should be overwritten or written as <clm_name>_Decoded}
}
\description{
Unicode hex value _hx[0-9A-F]{2}_ columns are converted into non alpha numeric text
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
decode_clms
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
