% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_decoding.R
\name{decode_special_char}
\alias{decode_special_char}
\title{Decodes unicode hex value}
\usage{
decode_special_char(char_vector)
}
\arguments{
\item{char_vector}{character vector that will be decoded}
}
\description{
Unicode hex value _hx[0-9A-F]{2}_ is converted into non alpha numeric text
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
decode_special_char
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
