% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_impres.R
\name{default_ckpt_pairs}
\alias{default_ckpt_pairs}
\title{Get the default gene pairs for the impres calculation}
\usage{
default_ckpt_pairs()
}
\value{
Returns data frame/table with 2 columns: sample ID & IMPRES score
}
\description{
Get the default gene pairs for the impres calculation
}
