% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackable_functions.R
\name{drop_low_columns}
\alias{drop_low_columns}
\title{Drop my_dt columns are below a certain percentile in value.}
\usage{
drop_low_columns(my_dt = NULL, col_names = NULL, col_value = median,
  my_summary = "Drop low expressing columns.",
  sample_key = get_default_sample_key(), percentile_cutoff = 10,
  readme_path = NULL)
}
\arguments{
\item{my_dt}{data.table input}

\item{col_names}{Vector of character strings to name the columns that will have this 
operation performed on them.  Uses \code{\link{operatable_columns}}}

\item{col_value}{Function that describes how column values are computed}

\item{my_summary}{Character string to change the default comment that will be appended 
to my_dt.}

\item{sample_key}{Character string to specify the column that is the sample key. This 
column will not be operated upon.}

\item{percentile_cutoff}{Numeric percentile.  Columns below this percentile will be 
dropped}

\item{readme_path}{Optional path to which the comments will be appended.}
}
\description{
Specifies a function to calculate the value of the column and a percentile, below which 
  samples will be dropped.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
drop_low_columns
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
