% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_fxns.R
\name{dXX_index}
\alias{dXX_index}
\title{dXX_index}
\usage{
dXX_index(my_counts, my_fraction, should_screen_counts = TRUE)
}
\arguments{
\item{my_counts}{vector of postive integers}

\item{should_screen_counts}{Boolean to indicate if the coutns should be screened
for valid data.  Set to false if the data has already been checked by another function.}
}
\value{
Typically called d50_index. Computes the diversity index dXX, where 
XX is a specified fraction.
}
\description{
From vdjtools:
"The estimate equals to 1 - n / N, where n is the minimum number of clonotypes 
accounting for at least XX% of the total reads and code N is the total number 
of clonotypes in the sample. Computes diversity index that equals to one minus 
the minimum fraction of clonotypes accounting for at least 50% of the total 
reads."
}
\seealso{
Other diversity: 
\code{\link{chao1}()},
\code{\link{evenness}()},
\code{\link{get_corrected_entropy_rdata_2_4096_path}()},
\code{\link{get_corrected_entropy_rdata_4097_65536_path}()},
\code{\link{get_corrected_entropy_rdata_ab1025_65K_ent1_8_path}()},
\code{\link{get_corrected_entropy_rdata_ab8_1024_ent1_8_path}()},
\code{\link{has_sufficient_abundance_for_entropy}()},
\code{\link{inv_simpson}()},
\code{\link{predict_true_entropy_from_counts}()},
\code{\link{predict_true_entropy_from_diversity}()},
\code{\link{screen_counts}()},
\code{\link{shannon_entropy}()}
}
\concept{diversity}
