% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_decoding.R
\name{encode_special_char}
\alias{encode_special_char}
\title{encodes non alpha numeric or underscore characters into hex.}
\usage{
encode_special_char(char_vector)
}
\arguments{
\item{char_vector}{character vector that will be encoded}
}
\description{
Non alpha numeric text (eg [a-zA-Z0-9_]) is converted to its unicode hex value
_hx[0-9A-F]{2}_
}
\details{
encoding_decoding
there are a lot of characters that aren't friendly for data.frame,/data.table
  operations as well as modeling and other functions.  Functions here will 
  encode r decode those values so they don't need so much manual intervention
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
encode_special_char
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
