% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_fxns.R
\name{evenness}
\alias{evenness}
\title{evenness}
\usage{
evenness(my_counts, should_screen_counts = TRUE)
}
\arguments{
\item{my_counts}{vector of postive integers}
}
\value{
Returns \code{shannon_entropy(my_counts)/log(sum(!is.na(my_counts)))} 
of the non-zero counts with NA removed.
}
\description{
evenness
}
\seealso{
Other diversity: 
\code{\link{chao1}()},
\code{\link{dXX_index}()},
\code{\link{get_corrected_entropy_rdata_2_4096_path}()},
\code{\link{get_corrected_entropy_rdata_4097_65536_path}()},
\code{\link{get_corrected_entropy_rdata_ab1025_65K_ent1_8_path}()},
\code{\link{get_corrected_entropy_rdata_ab8_1024_ent1_8_path}()},
\code{\link{has_sufficient_abundance_for_entropy}()},
\code{\link{inv_simpson}()},
\code{\link{predict_true_entropy_from_counts}()},
\code{\link{predict_true_entropy_from_diversity}()},
\code{\link{screen_counts}()},
\code{\link{shannon_entropy}()}
}
\concept{diversity}
