% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mixcr_analyze_shotgun_reports.R
\name{extract_mixcr_analyze_shotgun_reports}
\alias{extract_mixcr_analyze_shotgun_reports}
\title{Pull the numbers out of the reports from 'mixcr analyze shotgun'}
\usage{
extract_mixcr_analyze_shotgun_reports(input_paths)
}
\arguments{
\item{input_paths}{Vector of paths to the mixcr report files}
}
\value{
Returns data.table with samples by row and mixcr stats by column.
}
\description{
Pull the numbers out of the reports from 'mixcr analyze shotgun'
}
