% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_low_variance_clms.R
\name{find_low_variance_clms}
\alias{find_low_variance_clms}
\title{Returns column names of those whose most common factor is used by more than \code{max_fraction_largest} of the samples.}
\usage{
find_low_variance_clms(
  my_df,
  max_fraction_largest = max_fraction_largest,
  my_key = get_default_sample_key()
)
}
\arguments{
\item{my_df}{\code{data.frame} to search}

\item{max_fraction_largest}{number to indicate the max proportion the top fraction can take up before being reported.}

\item{my_key}{string to specify key column of the \code{data.frame}}
}
\description{
Checks all columns of \code{my_df} except \code{my_key} and checks if the most common factor is more than \code{max_fraction_largest}.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
find_low_variance_clms
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
