% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_similar_clms.R
\name{find_similar_clms}
\alias{find_similar_clms}
\title{Finds columns that have a high degree of correlation with this_clm}
\usage{
find_similar_clms(
  this_clm,
  my_df,
  test_clms = NULL,
  acceptable_similarity = 0.9,
  return_boolean = T,
  my_key = get_default_sample_key(),
  corr_method = "spearman"
)
}
\arguments{
\item{this_clm}{string to specify column that will be compared to the others}

\item{my_df}{\code{data.frame} to search}

\item{test_clms}{character vector of columns to test. If left blank all, but \code{this_clm} and \code{my_key} will be checked.}

\item{acceptable_similarity}{number to indicate the max absolute rho value that will be allowed with other columns.}

\item{return_boolean}{boolean to specify if a boolean is returned or the \code{found_clms}}

\item{my_key}{string to specify key column of the \code{data.frame}}

\item{corr_method}{string to specify the test used for the correlation. Passed to \code{method} arg of \code{\link{stats::cor.test}}}
}
\description{
Takes \code{this_clms} and compares it to all other columns of \code{my_df} using \code{\link{stats::cor.test}}.  If the absolute value of that row is greater than \code{acceptable_similarity}, then the either a boolean is returned true or the similar columns are returned.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
find_similar_clms
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
