% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{format_gene_set_name}
\alias{format_gene_set_name}
\title{Swaps hyphens for underscores}
\usage{
format_gene_set_name(my_string)
}
\arguments{
\item{my_string}{String from which to swap hyphens}
}
\description{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
format_gene_set_name
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
