% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{format_imported_annotation}
\alias{format_imported_annotation}
\title{Tabs out an imported annotations to isolate them visually.}
\usage{
format_imported_annotation(my_annotation)
}
\arguments{
\item{annotation}{Character vector of imported annotation.}

\item{indentation}{String of ' ' to insert before each line of the annotation to tab it over.}
}
\value{
The modified character vector.
}
\description{
Tabs out an imported annotations to isolate them visually.
}
