% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_central_clone.R
\name{generate_kmeans_cluster_list}
\alias{generate_kmeans_cluster_list}
\title{Obtain a list of clusters with their elements from data matrix sent}
\usage{
generate_kmeans_cluster_list(
  elements_df,
  cluster_id_width = NA,
  max_clusters = nrow(elements_df) - 1,
  min_clusters = 2,
  my_seed = NA
)
}
\arguments{
\item{elements_df}{Data.frame or matrix of data to be clustered as samples x elements ( i.e. features )}

\item{cluster_id_width}{The number of characters to include in cluster group and individual cluster id's ( will be used to left-pad cluster numbers with leading 0's )}

\item{max_clusters}{Integer number representing largest number of clusters to split data into. Must be less than number of rows in matrix.}

\item{min_clusters}{Integer number representing smallest number of clusters to split data into}

\item{my_seed}{The seed key to use before each call to kmeans method so each run can be reproduced}

\item{feature_df}{Data.frame with rows to be compared pair-wise}

\item{my_core_number}{Integer number representing the number of parallel processes to use for mhorn calculations}
}
\value{
Returns lists of clusters named as cluster_#clusters-in-run_cluster#-from-run = c(cluster_variable_names). Element names in clusters are sorted alphabetically.

Returns a correlation matrix in data.frame format
}
\description{
Runs kmeans method requesting each number of clusters from min_clusters:max_clusters

Compare each pair of rows in the feature_df using Morisita Horn method and return a correlation matrix ( as data.frame )
}
