% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{get_annotation_indention}
\alias{get_annotation_indention}
\title{Get the indentation string length that should be used for annotations.}
\usage{
get_annotation_indention()
}
\value{
Returns the default annotation indentaion string length.
}
\description{
Get the indentation string length that should be used for annotations.
}
