% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{get_annotation_path}
\alias{get_annotation_path}
\title{Get the path to the annotation file.}
\usage{
get_annotation_path(my_path)
}
\arguments{
\item{my_path}{The path to a folder or a file.}
}
\value{
The path to the found annotation file.
}
\description{
Takes path to a folder or a file in a folder. Finds THE '.note.' file in the folder and returns the path.
Assumes there is only one '.note.' file in each folder.
}
