% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{get_annotation_width}
\alias{get_annotation_width}
\title{Get the annotation width that should be used for annotations.}
\usage{
get_annotation_width()
}
\value{
Returns the default annotation width.
}
\description{
Get the annotation width that should be used for annotations.
}
