% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_navigation.R
\name{get_biomart_hsa_ucsc_path}
\alias{get_biomart_hsa_ucsc_path}
\title{get_biomart_hsa_ucsc_path}
\usage{
get_biomart_hsa_ucsc_path()
}
\arguments{
\item{none}{}
}
\value{
A path to the rds file.
}
\description{
Returns the path a saved mart
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
get_biomart_hsa_ucsc_path
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
\concept{mart}
