% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_navigation.R
\name{get_corrected_entropy_rdata_2_4096_path}
\alias{get_corrected_entropy_rdata_2_4096_path}
\title{get_corrected_entropy_rdata_2_4096_path}
\usage{
get_corrected_entropy_rdata_2_4096_path()
}
\value{
A path to the RData file
}
\description{
Returns the path to the corrected entropy model for abundances in the range of 2-4096.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
get_corrected_entropy_rdata_2_4096_path
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
\seealso{
Other diversity: 
\code{\link{chao1}()},
\code{\link{dXX_index}()},
\code{\link{evenness}()},
\code{\link{get_corrected_entropy_rdata_4097_65536_path}()},
\code{\link{get_corrected_entropy_rdata_ab1025_65K_ent1_8_path}()},
\code{\link{get_corrected_entropy_rdata_ab8_1024_ent1_8_path}()},
\code{\link{has_sufficient_abundance_for_entropy}()},
\code{\link{inv_simpson}()},
\code{\link{predict_true_entropy_from_counts}()},
\code{\link{predict_true_entropy_from_diversity}()},
\code{\link{screen_counts}()},
\code{\link{shannon_entropy}()}
}
\concept{diversity}
