% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackable_functions.R
\name{get_default_sample_key}
\alias{get_default_sample_key}
\title{Return the default f the column name for the default sample key}
\usage{
get_default_sample_key()
}
\arguments{
\item{none}{}
}
\description{
Created to access the default sample key column. unless otherwise specifed this is 
the column that will be assumed to contain the sample names
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
get_default_sample_key
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
