% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_navigation.R
\name{get_features_from_gmt_file}
\alias{get_features_from_gmt_file}
\title{get_features_from_gmt_file}
\usage{
get_features_from_gmt_file(gmt_path)
}
\arguments{
\item{gmt_path}{A string path to a gmt text file.}
}
\value{
A vector strings which includes the gene signature names
}
\description{
Just grabs the names of all of the gene sigs from a gmt file.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
get_features_from_gmt_file
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
\seealso{
Other gene_signatures: 
\code{\link{get_gene_set_entrez_human_bgvlab_path}()},
\code{\link{get_gene_set_entrez_human_c2_path}()},
\code{\link{get_gene_set_entrez_human_cta_path}()},
\code{\link{get_gene_set_entrez_human_kegg_path}()},
\code{\link{get_gene_set_hgnc_human_cta_path}()},
\code{\link{get_gene_set_hgnc_human_kegg_path}()}
}
\concept{gene_signatures}
