% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_model_data_et_al.R
\name{get_features_from_ranks}
\alias{get_features_from_ranks}
\title{Return vector of features names from a data.frame with rank values.}
\usage{
get_features_from_ranks(
  ranks_feature_path,
  ranks_count = NULL,
  ranks_clm = NULL
)
}
\arguments{
\item{ranks_feature_path}{Path to data matrix with feature names in one column and ranks in another.}

\item{ranks_count}{Integer to indicate how many features to take from the column ranks, \code{NULL} will return all the features ordered by rank.}

\item{ranks_clm}{String to indicate which column contains the column ranks. \code{NULL} will return the second column.}
}
\description{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
get_features_from_ranks
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
