% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_navigation.R
\name{get_gene_set_entrez_human_cta_path}
\alias{get_gene_set_entrez_human_cta_path}
\title{get_gene_set_entrez_human_cta_path}
\usage{
get_gene_set_entrez_human_cta_path()
}
\arguments{
\item{none}{}
}
\value{
A path to the gmt file.
}
\description{
Returns the path to the entrez_ids, human, CTA signatures in the package library.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
get_gene_set_entrez_human_cta_path
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
\seealso{
Other gene_signatures: 
\code{\link{get_features_from_gmt_file}()},
\code{\link{get_gene_set_entrez_human_bgvlab_path}()},
\code{\link{get_gene_set_entrez_human_c2_path}()},
\code{\link{get_gene_set_entrez_human_kegg_path}()},
<<<<<<< HEAD
=======
\code{\link{get_gene_set_hgnc_human_bgvlab_path}()},
\code{\link{get_gene_set_hgnc_human_c2cp_path}()},
>>>>>>> 0f15335346e040e4d7a17c1429b75222bde6c4cf
\code{\link{get_gene_set_hgnc_human_cta_path}()},
\code{\link{get_gene_set_hgnc_human_kegg_path}()}
}
\concept{gene_signatures}
