% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{get_gene_sigs_clustered_by_category}
\alias{get_gene_sigs_clustered_by_category}
\title{Gets the name of the immune gene signatures ordered by cell types.}
\usage{
get_gene_sigs_clustered_by_category()
}
\description{
Gets the name of the immune gene signatures ordered by cell types.
}
