% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{get_ipres_gene_sigs}
\alias{get_ipres_gene_sigs}
\title{Returns the names of the IPRES gene signatures.}
\usage{
get_ipres_gene_sigs()
}
\description{
Returns the names of the IPRES gene signatures.
}
