% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{get_note_extension}
\alias{get_note_extension}
\title{Gets the note extention from the global environment.}
\usage{
get_note_extension()
}
\arguments{
\item{None}{}
}
\value{
Returns the string that should be on the end of annotation files.
}
\description{
Gets the note extention from the global environment.
}
