% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_central_clone.R
\name{get_ranks_from_df}
\alias{get_ranks_from_df}
\title{function dependencies:}
\usage{
get_ranks_from_df(ranked_df, rank_clm = NA)
}
\description{
find_central_elements_by_cluster
    get_ranks_from_df
    compare_via_mhorn
    isolate_central_cluster_elements
        pca_central_element
        correlation_central_element
    cluster_list_to_df
    write_list_as_gmt
    write_central_elements_table
    write_ranked_central_elements_table
    pca_variance_plot
    cluster_group_plot
    pc1_vs_pc2_plot
    
group_clusters_by_size - useful for looking at results
collapse_identical_clusters - useful for looking at results
print_cluster_counts: takes output from collapse_identical_clusters
}
