% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{get_rho}
\alias{get_rho}
\title{Run cor.test and passes results}
\usage{
get_rho(vector_a, vector_b, my_method = "pearson")
}
\arguments{
\item{vector_a}{Numeric vector to correlate with vector_b}

\item{vector_b}{Numeric vector to correlate with vector_a}

\item{my_method}{string to pass to cor.test: 'pearson' or 'spearman'}
}
\description{
Run cor.test and passes results
}
