% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Get ggplot colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Integer for the number of colors to return}
}
\description{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
gg_color_hue
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
