% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Not in}
\usage{
... \%ni\% NA
}
\description{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
%ni%
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
