% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_central_clone.R
\name{group_clusters_by_size}
\alias{group_clusters_by_size}
\title{Group a list of clusters by size}
\usage{
group_clusters_by_size(
  c_list,
  min_size = 1L,
  max_size = NA,
  include_empty_groups = F
)
}
\arguments{
\item{c_list}{Named list of clusters - names will be preserved in groupings}

\item{min_size}{Integer number representing smallest cluster size to keep}

\item{max_size}{Integer number representing largest cluster size to keep}

\item{include_empty_groups}{Boolean whether or not to include cluster sizes in return value for which there are no representatives}
}
\value{
Returns lists of clusters ( with their elements ) grouped by cluster size named as clusters_of_x where x = cluster size.
}
\description{
Takes a list of clusters and groups them into named list by size
}
