% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_fxns.R
\name{has_sufficient_abundance_for_entropy}
\alias{has_sufficient_abundance_for_entropy}
\title{has_sufficient_abundance_for_entropy}
\usage{
has_sufficient_abundance_for_entropy(my_abundance, measured_entropy)
}
\arguments{
\item{my_abundance}{Integer of the total abundance of the counts}

\item{measured_entropy}{Measured entropy of the sample.  Value should be close to the 
range of (0.975-8.1).}
}
\value{
Boolean of whether this abundance is enough to just use the measured entropy
}
\description{
Over the range of TCGA TCR entropies (0.975-8.1) we calculated the minimum number 
of reads needed to get within the 95th CI of the entropy at 100 billion reads. 
See Optimize_Diversity_Metrics_CRSV1371 project file, plot_fraction_entropy.R.
}
\seealso{
Other diversity: \code{\link{chao1}},
  \code{\link{dXX_index}}, \code{\link{evenness}},
  \code{\link{get_corrected_entropy_rdata_2_4096_path}},
  \code{\link{get_corrected_entropy_rdata_4097_65536_path}},
  \code{\link{get_corrected_entropy_rdata_ab1025_65K_ent1_8_path}},
  \code{\link{get_corrected_entropy_rdata_ab8_1024_ent1_8_path}},
  \code{\link{inv_simpson}},
  \code{\link{predict_true_entropy_from_counts}},
  \code{\link{predict_true_entropy_from_diversity}},
  \code{\link{screen_counts}},
  \code{\link{shannon_entropy}}
}
\concept{diversity}
