% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{immune_gene_set_groupings}
\alias{immune_gene_set_groupings}
\title{Lists the immune gene signatures grouped by cell type.}
\usage{
immune_gene_set_groupings()
}
\description{
Lists the immune gene signatures grouped by cell type.
}
