% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{import_gmt_as_list}
\alias{import_gmt_as_list}
\title{Imports a gmt file as a list.}
\usage{
import_gmt_as_list(gmt_file_path)
}
\arguments{
\item{gmt_file_path}{path to gmt file}
}
\description{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
import_gmt_as_list
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
