% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{import_model_data}
\alias{import_model_data}
\title{Perform common steps needed for importing model data}
\usage{
import_model_data(
  ranks_file_pattern = "_ranks.tsv$",
  ranks_clm = NULL,
  ranks_count = NULL,
  merge_file_patterns = NULL,
  set_clm = "Set",
  sample_key = "Run_ID",
  indep_vars = NULL,
  dep_vars = NULL
)
}
\arguments{
\item{ranks_file_pattern}{String to use to look for a ranking of features}

\item{ranks_clm}{String to indicate which column contains the column ranks}

\item{ranks_count}{Integer to indicate houw many features to take from the column ranks}

\item{merge_file_patterns}{Character vector of file patterns to look for and merge to make the final output matrices}

\item{set_clm}{Column name to indicate which samples are in "Validation" vs "Discovery" sets}

\item{sample_key}{Column nameto use as the sample key to merge all matrices}

\item{indep_vars}{Character string of column names for the independent variables}

\item{dep_vars}{Character string of column names for the dependent variables}
}
\description{
All models will start with a matrix of ranked features, and the feature
values of the features. This fxn orders the features, pulls out the top x and
outputs disc and validation dataset matrices with the correct indep and dep
variables. This function has the option to specify certain features or to pull
the top x feature from a columns ranks file.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
import_model_data
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
