% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_fxns.R
\name{inv_simpson}
\alias{inv_simpson}
\title{inv_simpson}
\usage{
inv_simpson(my_counts, should_screen_counts = TRUE)
}
\arguments{
\item{my_counts}{vector of postive integers}

\item{should_screen_counts}{Boolean to indicate if the coutns should be screened
for valid data.  Set to false if the data has already been checked by another function.}
}
\value{
Returns \code{vegan::diversity(my_counts, index = "invsimpson")} of the 
non-zero counts with NA removed.
}
\description{
inv_simpson
}
\seealso{
Other diversity: \code{\link{chao1}},
  \code{\link{dXX_index}}, \code{\link{evenness}},
  \code{\link{get_corrected_entropy_rdata_2_4096_path}},
  \code{\link{get_corrected_entropy_rdata_4097_65536_path}},
  \code{\link{get_corrected_entropy_rdata_ab1025_65K_ent1_8_path}},
  \code{\link{get_corrected_entropy_rdata_ab8_1024_ent1_8_path}},
  \code{\link{has_sufficient_abundance_for_entropy}},
  \code{\link{predict_true_entropy_from_counts}},
  \code{\link{predict_true_entropy_from_diversity}},
  \code{\link{screen_counts}},
  \code{\link{shannon_entropy}}
}
\concept{diversity}
