% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{is_not_null}
\alias{is_not_null}
\title{Is not null}
\usage{
is_not_null(my_vector)
}
\arguments{
\item{my_vector}{A vector}
}
\value{
TRUE or FALSE for each item in the vector
}
\description{
Negates is.null.  Just reads better than \code{\link{!is.null}}.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
is_not_null
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
