% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{is_path_to_file}
\alias{is_path_to_file}
\title{Determines whether my_path goes to a file.}
\usage{
is_path_to_file(my_path, file_endings = c(".pdf", ".txt", ".R", ".xlsx",
  ".xlx", ".doc", ".tsv", ".csv"))
}
\arguments{
\item{my_path}{A path}

\item{file_endings}{character vector of file endings}
}
\value{
Boolean on whether the path ends in one of the file endings
}
\description{
Checks if my_path ends in one of the file_endings
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
is_path_to_file
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
