% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackable_functions.R
\name{log_transform_plus}
\alias{log_transform_plus}
\title{Does a log transformation of the data plus a specified value.}
\usage{
log_transform_plus(
  my_dt = NULL,
  add_value = 1,
  base = 2,
  col_names = NULL,
  my_summary = NULL,
  sample_key = get_default_sample_key(),
  readme_path = NULL
)
}
\arguments{
\item{my_dt}{data.table input}

\item{add_value}{Numeric value that will be added prior to the log transform.}

\item{base}{Numberic value providing the base for the log transform 
See \code{log} 'base.}

\item{col_names}{Vector of character strings to name the columns that will have this 
operation performed on them.  Uses \code{\link{operatable_columns}}}

\item{my_summary}{Character string to change the default comment that will be appended 
to my_dt.}

\item{sample_key}{Character string to specify the column that is the sample key. This 
column will not be operated upon.}

\item{readme_path}{Optional path to which the comments will be appended.}
}
\description{
This is a wrapper for \code{\link{log}} that also allows the adding of a value prior 
  to the transform.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
log_transform_plus
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
