% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{make_intro_text}
\alias{make_intro_text}
\title{Formats intro text for stackable functions}
\usage{
make_intro_text(function_name, my_summary = NULL)
}
\arguments{
\item{function_name}{Character name of function}

\item{my_summary}{Summary text to add to function intro text}
}
\description{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
make_intro_text
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
