% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{minimal_decimal_places_ordered}
\alias{minimal_decimal_places_ordered}
\title{Reduce decimal places needed}
\usage{
minimal_decimal_places_ordered(numeric_vector, min_decimal = 0)
}
\arguments{
\item{numeric_vector}{Values to represent}

\item{min_decimal}{Integer to indicate the number of folds in which to split the data}
}
\description{
Reduce the number of decimals down to the minimal needed to preserve the same number of levels in the original data. Returns data as factors to represent values in the same order they were given. 
This is handy for whne you
}
