% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_differential_expression.R
\name{model_differential_expression}
\alias{model_differential_expression}
\title{Does a differential gene expression analysis.}
\usage{
model_differential_expression(
  my_dt = NULL,
  analysis_method = "DESeq2",
  base_file_name = NULL,
  base_title = NULL,
  core_number = round(detectCores()/2),
  deseq2_results_cooksCutoff = NULL,
  deseq2_results_independentFiltering = TRUE,
  design_var = NULL,
  gene_expression_cols = NULL,
  gmt_file_fdr_cutoffs = c(0.2, 0.05),
<<<<<<< HEAD
  gmt_file_pvalue_cutoffs = c(0.01),
=======
  gmt_file_pvalue_cutoffs = c(0.05),
>>>>>>> 0f15335346e040e4d7a17c1429b75222bde6c4cf
  gmt_ref = "Gene signature created from custom analysis.",
  imported_annotation = NULL,
  my_grouping = NULL,
  output_dir,
  sample_key_col = "Sample_ID"
)
}
\arguments{
\item{analysis_method}{String option out of the following choices:  DESeq2, DESeq, SAM, or edgeR. Indicating which method
should be used to do the analysis.}

\item{base_file_name}{String to specify the file name.}

\item{core_number}{Integer to indicate the numboer of cores that should be used.}

\item{deseq2_results_cooksCutoff}{Set to \code{Inf} or \code{FALSE} to disable the resetting of p-values to \code{NA}. 
Gets passed to \code{\link{DESeq2::results}} 'independentFiltering' argument}

\item{deseq2_results_independentFiltering}{Gets passed to \code{\link{DESeq2::results}} 'cooksCutoff' argument}

\item{design_var}{This needs to be setup to allow more complex comparisons and and matching for samples.}

\item{gene_expression_cols}{Character vector with the names of the columns with genes in them.}

\item{gmt_file_fdr_cutoffs}{Numeric vector of cutoffs to use for the FDR significant values.  Two gene signtures
will be made of all the genes that have under the fdr pValue: one for up genes and one for down.}

\item{gmt_file_pvalue_cutoffs}{Numeric vector of cutoffs to use for the pValue significant values.  Two gene signtures
will be made of all the genes that have under the pValue: one for up genes and one for down.}

\item{gmt_ref}{String indicating what should go in the reference part of the gmt file.}

\item{imported_annotation}{Character vector to include what steps were done to the data prior to this analysis.  This module will 
add on to those steps.}

\item{my_grouping}{This string is the name of the column you want to use to split the data into groups.}

\item{output_dir}{Path to the output directory.  This will be calculated automatically if left blank.}

\item{sample_key_col}{String matching the name of the column that should be used to identify the unique sample identifiers.}
}
\value{
List containing several outputs contains: 
\enumerate{
  \item heatmap_pack - to be used by \code{\link{view_heatmap}}; 
  \item volcano_pack - to be used by \code{\link{view_volcano_plot}}; 
  \item gmt_id_path - path to a gmt file with genes represetned by ids (typically entrez);
  \item gmt_symbol_path - path to a gmt file with genes represented by symbols (typically hgnc);
  \item stats_path - path to stats of Gene_Name, Fold_Change, pValue, FDR_pValue
}
}
\description{
The purpose of \code{model_differential_expression} is to compare the raw read counts of gene expression data 
between different groups of samples to see if there is differential gene expression. This is different from model_gene_expression in that this one 
tries to incorporate the changes needed for the LCCC-Bioinformatics group to use. Dumping sam and edger here. Letting
old Deseq (ie not deseq2) go for a bit until it's needed.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
model_differential_expression
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~


This function utilizes one of either \code{\link{DESeq}} or \code{\link{DESeq2}} methods.  \code{\link{DESeq}} is recommended for single cell data.
}
\section{Writes}{

\itemize{
  \item stats file 
}
}

\section{Todos}{

\itemize{
  \item Fix to match by patient ids.
  \item Setup design_var and contrasts
}
}

\section{Limitations}{

\itemize{
  \item Can only match two samples at this point.
}
}

\section{See also}{

\itemize{
  \item \code{\link{view_heatmap}}
  \item \code{\link{view_volcano_plot}}
}
}

\concept{differential_expression, model}
