% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_regression.R
\name{model_regression}
\alias{model_regression}
\title{performs the specified regression on the data}
\usage{
model_regression(
  input_dt,
  indep_list,
  base_file_name = "regression_output",
  clear_readme = TRUE,
  combined_group_name = "All",
  dep_vars = NULL,
  dep_var_families = NULL,
  event_col = "OS_e",
  fdr_by_columns = NULL,
  fdr_by_columns_for_model_comp = NULL,
  fdr_method = "BH",
  inclusion_list = list(),
  model_comparison_list = NULL,
  model_function = function(dep_var = "", indep_vars = "NULL") {     paste0("glm(",
    dep_var, " ~ ", paste0(indep_vars, collapse = " + "), ", data = model_dt)") },
  my_grouping = NULL,
  output_dir = ".",
  time_col = "OS_d",
  write_files = TRUE
)
}
\arguments{
\item{input_dt}{A data.table that includes all columns of data needed to do the 
analysis:  \code{names(indep_list)}, dep_vars (for glm), \code{names(inclusion_list)},
event_col & time_col (for coxph), \code{names(unique(unlist(model_comparison_list))},
and my_grouping.}

\item{indep_list}{Required named list of column names to use as the independent 
variable. Names of the list will be used to name the output stats.
Example:  \cr 
\code{my_indep_list = list( # no default \cr 
  TRA_Chao1 = c("TRA_Chao1"), \cr 
  TRB_Chao1 = c("TRB_Chao1"), \cr 
  TCR_Chao1 = c("TRA_Chao1", "TRB_Chao1") \cr 
)}}

\item{base_file_name}{Character string to prefix the names of the output files.}

\item{combined_group_name}{Character string to call the combinded groups catagory.}

\item{dep_vars}{Character vector containing the coumn names. Example: \cr 
\code{codemy_dep_vars = c("Age", "SNV_Log2_Neoantigens", "Indel_Log2_Neoantigens")}}

\item{dep_var_families}{This character vector should contain the names of the 
families to add to the model.  This isn't used for coxph, since the dependent
variable for coxph is survival.  Possible values here should be of the form: 
\code{'Gamma("identity")'} or \code{'gaussian'} and can include anything accepted 
\code{\link{glm}}.}

\item{event_col}{For coxph.  The name of the column from which to draw the event 
information.  The column should only contain integers of 1 and 0.  If specified, 
this column needs to be present in \code{input_dt}.}

\item{fdr_by_columns}{Passed to \code{\link{binfotron::calc_fdr}}. Optional 
character vector of column names to specify what should be group together in 
the fdr correction using the data.table "by" option.  Leaving this blank will 
just apply the correction to all of the data values. "Group" is a common value
to use here.}

\item{fdr_by_columns_for_model_comp}{Passed to \code{\link{binfotron::calc_fdr}}. 
Optional character vector of column names to specify what should be group together 
in the fdr correction of the model comparison stats using the data.table "by" 
option.  Leaving this blank will just apply the correction to all of the data 
values. "Group" is a common value to use here.}

\item{fdr_method}{Passed to \code{\link{binfotron::calc_fdr}}. String to tell 
what method to use to FDR correct. Must be one of the values in 
\code{stats::p.adjust.methods}.}

\item{inclusion_list}{List to specify the samples that should be kept.
For example \code{list(pathology_T_stage = c('T1', 'T2', 'T3'), is_asian = c(TRUE))} 
would drop samples in which the value for the column named 'pathology_T_stage' 
was not either 'T1', 'T2', or 'T3'. Samples must also have 'is_asian' equal 
to \code{TRUE}.  The names of this list should be column names for \code{input_dt}}

\item{model_comparison_list}{Optional named list of column names that should be used for a 
full and reduced model comparison.  Every group of coulms on this list will be 
run against each dep_vars indep_list combination.  The reduced model will only 
include the items on the list.  The full modle will include the independent 
varible(s) as well.  The names of this list will be what the model comparison 
will be called. The values of this list should be column names in \code{input_dt}.
Example: \cr 
\code{model_comparison_list = list( \cr 
  Age = c("Age"), \cr 
  Tissue = c("Tissue"), \cr 
  Combined = c("Age", "Tissue") \cr 
)}}

\item{model_function}{A function to return the model.  Important to set 
\code{data = model_dt} in the function.  Do not set glm family.  This will be
added based on \code{dep_var_families}. See examples below.
glm example: \cr 
\code{model_function =function(dep_var = "", indep_vars = "NULL"){ \cr 
  paste0("glm(", dep_var, " ~ ",  paste0(indep_vars, collapse = " + "), ", data = model_dt)") \cr   
}} \cr 
coxph example: \cr 
\code{model_function = function(dep_var = "", indep_vars = "NULL"){ \cr 
   paste0("coxph(Surv(", time_col,", ", event_col, ") ~ ",  paste0(indep_vars, collapse = " + "), ", data = model_dt)") \cr 
}} \cr}

\item{my_grouping}{This string is the name of the column you want to use to split 
the data into groups. If specified, this column needs to present in \code{input_dt}.}

\item{output_dir}{Path to the output directory. The parent directory to the path must exist.}

\item{time_col}{For coxph.  The name of the column from which to draw the time 
information.  If specified, this column needs to present in \code{input_dt}.}

\item{write_files}{Boolean on whether you wuold like to write the output files.}
}
\value{
List containing several outputs: 
\enumerate{
  \item stats - data.table with the results of the model output
  \item model_comp - data.table with the full vs reduced model comparisons if \code{model_comparison_list} is provided 
  \item readme - An output of the comparisons made.
}
}
\description{
The purpose of \code{model_regression} is to perform a regression on the data 
across the range of independant and dependant variables provided.  If m
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
model_regression
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~


This function utilizes one of either \code{\link{glm}} or \code{\link{coxph}} methods.
}
\section{Writes}{

\itemize{
  \item stats file 
  \item model_comp file if \code{model_comparison_list} is provided.
}
}

\section{Todos}{

\itemize{
  \item This is new so will need some debugging.
  \item Support GAMLSS and ability to determine its own family??
  \item Stats output sometimes has blank lines in it
}
}

\section{Limitations}{

\itemize{
  \item Haven't fixed to run with ordinals. 
  \item Haven't tried models with interactions with it yet
}
}

\concept{model}
