% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackable_functions.R
\name{normalize_rows_by_quartile}
\alias{normalize_rows_by_quartile}
\title{Normalize data values across rows by quartile}
\usage{
normalize_rows_by_quartile(
  my_dt = NULL,
  col_names = NULL,
  my_summary = "Upper-quartile normalizing data across rows.",
  norm_factor = 1000,
  percentile = 75,
  sample_key = get_default_sample_key(),
  readme_path = NULL
)
}
\arguments{
\item{my_dt}{data.table input}

\item{col_names}{Vector of character strings to name the columns that will have this 
operation performed on them.  Uses \code{\link{operatable_columns}}}

\item{my_summary}{Character string to change the default comment that will be appended 
to my_dt.}

\item{norm_factor}{Number indicating the value to which the quartile is normalized}

\item{percentile}{Percentile to which the data will be normalized}

\item{sample_key}{Character string to specify the column that is the sample key. This 
column will not be operated upon.}

\item{readme_path}{Optional path to which the comments will be appended.}
}
\description{
Normalizes across rows to a certain quartile. Basically wraps \code{\link{quantile}}
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
normalize_rows_by_quartile
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
