% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{operatable_columns}
\alias{operatable_columns}
\title{Determines what column names can be operated upon.}
\usage{
operatable_columns(col_names = NULL, my_dt, acceptable_classes = NULL,
  sample_key = "Sample_ID")
}
\arguments{
\item{col_names}{Vector of character strings to name the columns that will have this 
operation performed on them.}

\item{my_dt}{data.table input}

\item{acceptable_classes}{Character vector of classes that are acceptable to perform 
whatever operation.  col_names not of this class will not be returned.}

\item{sample_key}{Character string to specify the column that is the sample key. This 
column will not be operated upon.}
}
\description{
Uses column classes and sample_key to determine what columns can be operated upon.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
operatable_columns
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
