% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_central_clone.R
\name{pc1_vs_pc2_plot}
\alias{pc1_vs_pc2_plot}
\title{Saves to file a plot of pc1 vs pc2 from a pca analysis}
\usage{
pc1_vs_pc2_plot(scores_df, output_path = NA, my_width = 8, my_height = 8)
}
\arguments{
\item{scores_df}{Dataframe with elements x principle components. Expects at least a PC1 and PC2 column}

\item{output_path}{Character vector with fully qualified path to which plot will be saved}

\item{my_width}{One-length numeric vector with width of output plot ( in inches )}

\item{my_height}{One-length numeric vector with height of output plot ( in inches )}
}
\value{
Returns the ggplot object
}
\description{
Saves to file a scatterplot showing PC1 vs PC2
}
