% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_central_clone.R
\name{pca_variance_plot}
\alias{pca_variance_plot}
\title{Saves to file a cumulative variance plot for principle components of a PCA analysis}
\usage{
pca_variance_plot(pca_results, output_path = NA, my_width = 8, my_height = 8)
}
\arguments{
\item{pca_results}{A prcomp object}

\item{output_path}{One-length character vector with fully qualified path to output file}

\item{my_width}{One-length numeric vector for the width of the output plot}

\item{my_height}{One-length numeric vector for the height of the output plot}
}
\value{
Returns the ggplot object
}
\description{
Saves to file a cumulative variance plot for principle components of a PCA analysis with optional coloring by cumulative percent variance explained
}
