% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_features.R
\name{plot_cumulative_features}
\alias{plot_cumulative_features}
\title{Plots a line depicting the number of features ( or any categorical variable ) that are below the pvalue ( or any categorical variable )}
\usage{
plot_cumulative_features(
  plot_df,
  gg_add_on = theme_minimal(),
  lab_x = "P-Value ( Log10 )",
  lab_y = "Cumulative Feature Count ( Log10 )",
  max_features = NA,
  max_pvalue = NA,
  output_path = NA,
  plot_title = "Cumulative Features Below P-Value",
  pvalue_clm = "pValue",
  size_height = 5,
  size_width = 5.5
)
}
\arguments{
\item{plot_df}{Dataframe containing at a minimum pvalue data}

\item{gg_add_on}{ggplot elements to add on to plot. Defaults to \code{theme_minimal()}}

\item{lab_x}{String for labeling x axis}

\item{lab_y}{String for labeling y axis}

\item{max_features}{integer representing the maximum number of features to plot. Values above this will be dropped from the plot_df}

\item{max_pvalue}{double representing the highest pvalue to be plotted. Values above this will be dropped from the plot_df}

\item{output_path}{Output path for plot}

\item{plot_title}{String for plot title}

\item{pvalue_clm}{column name in plot_df corresponding to pvalues}

\item{size_height}{Numeric to specify plot height}

\item{size_width}{Numeric to specify plot width}
}
\value{
Returns ggplot object with graphed line and vertical lines at significance intervals
}
\description{
Intended to depict how many features are below pvalues with vertical lines representing the typical cutoffs for significance ( .05, .01, .001 )
}
