% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_fxns.R
\name{predict_true_entropy_from_diversity}
\alias{predict_true_entropy_from_diversity}
\title{predict_true_entropy_from_diversity}
\usage{
predict_true_entropy_from_diversity(my_abundance, my_richness, my_d25,
  my_inv_simpson, my_chao1, my_shannon_entropy, my_evenness,
  min_abundance = 2, should_screen_counts = TRUE)
}
\arguments{
\item{min_abundance}{Integer to indicate minimun number of counts needed (ie. \code{sum(my_counts)})
to get a valid prediction.}

\item{should_screen_counts}{Boolean to indicate if the coutns should be screened
for valid data.  Set to false if the data has already been checked by another function.}

\item{my_counts}{vector of postive integers}

\item{max_abundance}{Integer over which the function will return shannon entropy.}
}
\value{
Modeled prediction of diversity
}
\description{
Uses an elastic net model to predict true entropy.  Two models are used to cover
different abundnce ranges.  2-1024 is one model 1025-65536 is another.  Over 2^16
just returns the shannon entropy, which is pretty accurate over this amount. These
models were build around tcga tcr distributions that had an entropy range of 0.975 
to 8.1 if corrected entropy is outside of this range then the data should be taken 
with caution.
}
\seealso{
Other diversity: \code{\link{chao1}},
  \code{\link{dXX_index}}, \code{\link{evenness}},
  \code{\link{get_corrected_entropy_rdata_2_4096_path}},
  \code{\link{get_corrected_entropy_rdata_4097_65536_path}},
  \code{\link{get_corrected_entropy_rdata_ab1025_65K_ent1_8_path}},
  \code{\link{get_corrected_entropy_rdata_ab8_1024_ent1_8_path}},
  \code{\link{has_sufficient_abundance_for_entropy}},
  \code{\link{inv_simpson}},
  \code{\link{predict_true_entropy_from_counts}},
  \code{\link{screen_counts}},
  \code{\link{shannon_entropy}}
}
\concept{diversity}
