% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_central_clone.R
\name{print_cluster_counts}
\alias{print_cluster_counts}
\title{Display cluster distribution within groups by cluster size}
\usage{
print_cluster_counts(
  unique_clusters_by_size,
  cluster_lengths = 1:max(mapply(function(x) length(x[[1]]), unique_clusters_by_size))
)
}
\arguments{
\item{unique_clusters_by_size}{List of cluster sizes, each size being a list of unique clusters as output by this packages' collapse_identical_clusters method}

\item{cluster_lengths}{Integer range of cluster sizes to include in output}
}
\value{
Returns list of requested cluster sizes with elements of unique clusters and stats on frequency
}
\description{
Prints out elements of unique clusters of each size requested with stats on frequency of occurrence
}
