% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackable_functions.R
\name{require_data_in_samples}
\alias{require_data_in_samples}
\title{Drop my_dt columns that aren't present in certain percentage_of_samples according
  to requirement function}
\usage{
require_data_in_samples(my_dt = NULL, col_names = NULL,
  my_summary = "Require data in smaples", percentage_of_samples = 75,
  requirement = function(x) {     sum(x > 0) },
  sample_key = get_default_sample_key(), readme_path = NULL)
}
\arguments{
\item{my_dt}{data.table input}

\item{col_names}{Vector of character strings to name the columns that will have this 
operation performed on them.  Uses \code{\link{operatable_columns}}}

\item{my_summary}{Character string to change the default comment that will be appended 
to my_dt.}

\item{percentage_of_samples}{Percentage of samples that will need to meet the requirement 
to keep the column}

\item{requirement}{Function by which to judge the column. For example the default 
(\code{function(x){sum(x>0)}}) requires that the value be greater than zero in 75 
percent of the samples.}

\item{sample_key}{Character string to specify the column that is the sample key. This 
column will not be operated upon.}

\item{readme_path}{Optional path to which the comments will be appended.}
}
\description{
Allows you drop columns that don't meet criteria for X% of samples
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
require_data_in_samples
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
