% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_annotation.R
\name{rjust}
\alias{rjust}
\title{Right justifies my_string.}
\usage{
rjust(my_string, filler = " ", total_char = get_annotation_width())
}
\arguments{
\item{my_string}{String to be left justified}

\item{filler}{String character to be used to space the text over.}

\item{total_char}{Integer of how many characters the text should be when done.}
}
\value{
Right justified text.
}
\description{
Adds filler characters to the left of my_string to have the text fill the line up
total_char characters.
}
