% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackable_functions.R
\name{scale_columns}
\alias{scale_columns}
\title{Scales data.table columns.}
\usage{
scale_columns(my_dt = NULL, col_names = NULL,
  my_summary = "Scale the data by column.",
  sample_key = get_default_sample_key(), should_scale = TRUE,
  should_center = TRUE, readme_path = NULL)
}
\arguments{
\item{my_dt}{data.table input}

\item{col_names}{Vector of character strings to name the columns that will have this 
operation performed on them.  Uses \code{\link{operatable_columns}}}

\item{my_summary}{Character string to change the default comment that will be appended 
to my_dt.}

\item{sample_key}{Character string to specify the column that is the sample key. This 
column will not be scaled.}

\item{should_scale}{Boolean to indicate whether the data should be scaled or not.
See \code{\link{scale}} 'scale' argument.}

\item{should_center}{Boolean to indicate whether the data should be centered or not. 
See \code{\link{scale}} 'center' argument.}

\item{readme_path}{Optional path to which the comments will be appended.}
}
\description{
A stackable wrapper for \code{\link{scale}} to turn it into a stackable transformation
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
scale_columns
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
