% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_fxns.R
\name{screen_counts}
\alias{screen_counts}
\title{screen_counts}
\usage{
screen_counts(my_counts)
}
\arguments{
\item{my_counts}{return valid counts}
}
\value{
Returns couns that are not NA and greater than zero.  Factors are 
converted to numerics correctly.
}
\description{
screen_counts
}
\seealso{
Other diversity: \code{\link{chao1}},
  \code{\link{dXX_index}}, \code{\link{evenness}},
  \code{\link{get_corrected_entropy_rdata_2_4096_path}},
  \code{\link{get_corrected_entropy_rdata_4097_65536_path}},
  \code{\link{get_corrected_entropy_rdata_ab1025_65K_ent1_8_path}},
  \code{\link{get_corrected_entropy_rdata_ab8_1024_ent1_8_path}},
  \code{\link{has_sufficient_abundance_for_entropy}},
  \code{\link{inv_simpson}},
  \code{\link{predict_true_entropy_from_counts}},
  \code{\link{predict_true_entropy_from_diversity}},
  \code{\link{shannon_entropy}}
}
\concept{diversity}
