% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_model_data_et_al.R
\name{serial_import}
\alias{serial_import}
\title{Import a vector fo file paths into one matrix.}
\usage{
serial_import(
  import_paths,
  df_filters = c(),
  sample_key = binfotron::get_default_sample_key(),
  import_clms = NULL,
  keep_all_samples = FALSE
)
}
\arguments{
\item{import_paths}{Vector of paths to import and merge}

\item{df_filters}{Character vector to be parsed in a loop using \code{dplyr::filter}}

\item{sample_key}{Column name to use as the sample key to merge all matrices}

\item{import_clms}{Vector of column names to import}

\item{keep_all_samples}{TRUE will keep all samples from all imported matrices. FALSE will only keep samples which are in all matrices}
}
\description{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
serial_import
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
