% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_scale.R
\name{soft_scale}
\alias{soft_scale}
\title{Winsorized scaling}
\usage{
soft_scale(input_data, cap = scale_cap)
}
\arguments{
\item{input_data}{Numeric vector to be scaled}

\item{cap}{Number indicating the number of sd that you'd like the value to be limited}
}
\description{
Replaces outliers beyond a certain number of standard deviations, ie cap, with a value at that sd.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
soft_scale
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
