% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{specify_decimal}
\alias{specify_decimal}
\title{String rounded to a specific decimal places}
\usage{
specify_decimal(my_number, number_of_digits)
}
\arguments{
\item{my_number}{Number that will be rounded}

\item{number_of_digits}{Integer of didgits to round to.}
}
\description{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
specify_decimal
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
