% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{t_test_ci}
\alias{t_test_ci}
\title{calculates mean & CI for residuals using t-distribution}
\usage{
t_test_ci(x)
}
\arguments{
\item{x}{Numeric vector to correlate with vector_b}
}
\description{
This CI method does not assume we know the sd and uses different t-distributions to get a better estimate of the distribution of the data.
Both t_test_ci and z_test_ci assume the data are normally distributed. It will usually be better to use the t_test_ci. This will equal the z_test_ci as N -> inf
}
