% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_volcano_plot.R
\name{view_volcano_plot}
\alias{view_volcano_plot}
\title{Create a volcano plot of analysis results.}
\usage{
view_volcano_plot(
  my_pack = NULL,
  base_file_name = NULL,
  base_title = NULL,
  colname_of_feature_groups = NULL,
  colname_of_feature_names = NULL,
  colname_of_fold_change = NULL,
  colname_of_pValue = NULL,
  colname_of_FDR = NULL,
  imported_annotation = NULL,
  my_dt = NULL,
  ordered_factors = NULL,
  output_dir = NULL,
  add_pdf_metadata = TRUE,
  axis.title.x.size = 12,
  axis.title.y.size = 16,
  FDR_label_threshhold = NULL,
  fdr_method = "BH",
  FDR_sig_alpha = 1,
  label_font_size = 2.5,
  label_pValue_line = TRUE,
  log_completion = FALSE,
  max_number_of_labels = 50,
  nonsig_label = NULL,
  num_top_groups_to_color = 6,
  n_to_label = NULL,
  output_subfolder = "XX_Volcano_Plot",
  pdf_bookmark_title = "02_Volcano_Plot",
  pdf_height = 7,
  pdf_width = 9,
  plot_by_FDR = FALSE,
  pValue_label_threshhold = NULL,
  pValue_line = NULL,
  sig1_cutoff = 0.05,
  sig1_label = NULL,
  sig1_type = "FDR",
  sig2_cutoff = 0.05,
  sig2_label = NULL,
  sig2_type = "pValue",
  strings_to_cut_from_labels = c("MOUSECYC_MM", "PWY", "WIKIPATHWAYS_MM", "PANTHER_MM",
    "BIOCARTA_MM", "WIKIPATHWAYS_MM", "KEGG_MM", "TF_MM", "REACTOME_MM", "TFACTS_MM",
    "PID_", "KEGG_", "BIOCARTA", "REACTOME_", "NABA_"),
  x_axis_label = NULL
)
}
\arguments{
\item{my_pack}{A named list for analysis modules to send parameters to the view function in a 
more compact fashion:
\itemize{
  \item base_file_name
  \item base_title
  \item colname_of_feature_groups
  \item colname_of_feature_names
  \item colname_of_fold_change
  \item colname_of_pValue
  \item colname_of_FDR
  \item imported_annotation
  \item my_dt
  \item ordered_factors
  \item output_dir
}
These list items may be overridden by specifying them individually.  They are described in 
more detail below.}

\item{base_file_name}{Overrides \code{my_pack$base_file_name}. String to specify the file name.}

\item{base_title}{Overrides \code{my_pack$base_title}. String to specify the title of the graph.}

\item{colname_of_feature_groups}{Overrides \code{my_pack$colname_of_feature_groups}. String indicating the name of the column to get the feature groups from.
If providing groups the points will be colored according to these groups and not significance.}

\item{colname_of_feature_names}{Overrides \code{my_pack$colname_of_feature_names}. String indicating the name of the column to pull the feature names from.}

\item{colname_of_fold_change}{Overrides \code{my_pack$colname_of_fold_change}. String indicating the name of the column to get the fold change from.}

\item{colname_of_pValue}{Overrides \code{my_pack$colname_of_pValue}. String indicating the name of the column to get the pValue from.}

\item{colname_of_FDR}{Overrides \code{my_pack$colname_of_FDR}. String indicating the name of the column to get the FDR from. If not provided the fdr_method 
will be used to calculate FDR values.  fdr_method should be set to "none" if no correction is desired.}

\item{imported_annotation}{Overrides \code{my_pack$my_annotation}. Annotations describe the 
steps taken during the analysis module and previous steps. This function will also add 
annotation steps to provide a full description of how the graph was generated.}

\item{my_dt}{Data table to provide the information for the graph.  It should have columns for pValue and fold_change 
as a minimum.}

\item{ordered_factors}{Overrides \code{my_pack$ordered_factors}. A character string of length 2 to tell the x axis what it should be labeled}

\item{output_dir}{Overrides \code{my_pack$output_dir}. Path to folder where the heatmap output should go.
Analysis modules will determine this automatically.}

\item{add_pdf_metadata}{Boolean indicating whether the pdf metadata should be chnaged.  Is pdftk present?}

\item{axis.title.x.size}{Numeric indicating the size of the x axis title.}

\item{axis.title.y.size}{Numeric indicating the size of the y axis title.}

\item{FDR_label_threshhold}{Number to specify threshold over which points will not be labeled.}

\item{fdr_method}{String indicting the type of fdr method to use if no colname_of_FDR is given.  See \code{\link{p.adjust}} 'method'.}

\item{FDR_sig_alpha}{numeric to indicate the alpha value of fdr significant points}

\item{label_font_size}{Numeric to indicate the size of the point labels.}

\item{label_pValue_line}{Boolean to determine whether the pValue line should be labeled.}

\item{log_completion}{Binary indicting whether the graph should be added to the table of completed analyses.}

\item{max_number_of_labels}{integer indicating the max number of points that will be labeled.}

\item{nonsig_label}{String to label legend for non-significant points.}

\item{num_top_groups_to_color}{Integer that tells how many of the top groups (using group argument) should be labeled and colored.}

\item{n_to_label}{Integer to specify how many points should be labeled.  More than 25 or so can get a bit cluttered. This argument supercedes FDR and pValue label threshholds. Set to 0 if you don't want labels}

\item{output_subfolder}{String that specifies what folder the output will go into in the main output directory.}

\item{pdf_bookmark_title}{String indicating the name of the bookmark that will be used when concatentaing the pdf's into a report}

\item{pdf_height}{Sets the height (inches) of the pdf output}

\item{pdf_width}{Sets the width (inches) of the pdf output}

\item{plot_by_FDR}{boolean to indicate if the y axis should be against the FDR_corrected pvalues}

\item{pValue_label_threshhold}{Number to specify threshold over which points will not be labeled.}

\item{pValue_line}{Number to specify where to draw a line to label a specific pValue. Usually set to 0.05 if used.}

\item{sig1_cutoff}{Numeric to indicate the upper bound for the most significant points.}

\item{sig1_label}{String to label legend for the most significant points.}

\item{sig1_type}{Either "FDR" or "pValue". Indicates what should be used to determine the most significant points.}

\item{sig2_cutoff}{Numeric to indicate the upper bound for the second most significant points.}

\item{sig2_label}{String to label legend for the second most significant points.}

\item{sig2_type}{Either "FDR" or "pValue". Indicates what should be used to determine the second-most significant points.}

\item{strings_to_cut_from_labels}{In certain gene group names a lot of the names can be very long and not very informative. This vector of character strings allows you to specify strings that will be cut out of the names. Leave off leading and trailing "_".}

\item{x_axis_label}{String to specify label for x axis.}
}
\value{
1) Volcano plot; 
  2) Volcano plot pdf;
}
\description{
Fold change and significance are used to produce a volcano plot of data for gene expression, 
DAVID analysis and gene set colloctions
}
\details{
\code{bgv_volcano_plot} points may be labeled based on being over a certain significance or just
by labeling a certain number to label.  Clusters of points may also be colored to show how the group is doing.
}
\section{Todos}{

\itemize{
  \item Debug for groups
}
}

\section{Future changes}{

\itemize{
  \item ...
}
}

\concept{view}
