% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_central_clone.R
\name{write_central_elements_table}
\alias{write_central_elements_table}
\title{Saves to file a dataframe of elements ( i.e. features ) x cluster groups}
\usage{
write_central_elements_table(
  central_elements,
  all_elements,
  output_path = file.path(getwd(), "central_elements_by_num_clusters.tsv"),
  return_output = F,
  cluster_id_width = NA
)
}
\arguments{
\item{central_elements}{Named list of central element from each cluster ( as returned by isolate_central_cluster_elements )}

\item{all_elements}{Character vector of element names containing, at a minimum, all central elements defined in first argument}

\item{output_path}{One-length character vector with fully qualified path to output file}

\item{return_output}{Boolean option to display final output}

\item{cluster_id_width}{The number of characters in cluster group id's and individual cluster id's}
}
\value{
Returns central elements data.frame as saved to file
}
\description{
Creates a file with one tab separated row per each element and one column for each number of clusters with 1's for the central elements of each cluster in each cluster group and "" everywhere else
}
