% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_central_clone.R
\name{write_list_as_gmt}
\alias{write_list_as_gmt}
\title{Saves to file list of element combinations and most central element from each grouping}
\usage{
write_list_as_gmt(
  cluster_members,
  central_elements,
  output_path = file.path(getwd(), "output_file.gmt.txt"),
  empty_cell = "",
  display_output = F
)
}
\arguments{
\item{cluster_members}{Named list of clusters ( names will be used as clusterid's )}

\item{central_elements}{Named list of central elements ( i.e. features ) from each cluster ( same length as cluster_members with matching names )}

\item{output_path}{One-length character vector with fully qualified path to output file}

\item{empty_cell}{Filler for empty cells. Typically '' or NA. Note there is a gene called 'NA'.  It's in the influenza genome but still...}

\item{display_output}{Boolean option to display which elements are in the final signatures.  For gene signatures this will typically be too much information to display}
}
\value{
Returns gene signature data as saved to file
}
\description{
Creates a file with one tab separated line per each cluster of elements formatted as central_element__clusterid, comment, full list of elements in this gene signature with NA's for extra spots to length of largest cluster
}
