% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{write_metadata}
\alias{write_metadata}
\title{Adds metadata to pdf}
\usage{
write_metadata(save_path, pdf_metadata)
}
\arguments{
\item{save_path}{Character string for directory}

\item{pdf_metadata}{Character vector of pdf metadata to add to pdf}
}
\description{
Adds metadata to pdf
}
