% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{z_test_ci}
\alias{z_test_ci}
\title{calculates mean & CI for residuals using z-score sd}
\usage{
z_test_ci(x, show_warning = T)
}
\arguments{
\item{x}{Numeric vector to correlate with vector_b}

\item{show_warning}{Boolean to indicate whether a warning be shown if there are too few samples to accurately use this method to get confidence intervals}
}
\description{
The z test assumes we know the sd ( ie we have a large number of samples n > 30). Both t_test_ci and z_test_ci assume the data are normally distributed. It will usually be better to use the t_test_ci. This will equal the t_test_ci as N -> inf
}
