% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{pvalue_nums}
\alias{pvalue_nums}
\title{Return string indicating significance level bin of p.value provided}
\usage{
pvalue_nums(pvalues, sig_levels = c(0.05, 0.01, 0.001))
}
\arguments{
\item{sig_levels}{vector of PValues to define our bin levels}

\item{p.value}{double representing the p.value to obtain significance level for}
}
\value{
Returns string representing the significance bin of this p.value
}
\description{
Method to determine significance level bin and return a formatted string. Duplicate of binfotron::pvalue_stars but returns values like '<= 0.01'  rather than '**'
}
