% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_model_data_et_al.R
\name{serial_import}
\alias{serial_import}
\title{Import a vector fo file paths into one matrix.}
\usage{
serial_import(
  import_paths,
  sample_key = binfotron::get_default_sample_key(),
  import_clms = NULL
)
}
\arguments{
\item{import_paths}{Vector of paths to import and merge}

\item{sample_key}{Column name to use as the sample key to merge all matrices}

\item{import_clms}{Vector of column names to import}
}
\description{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
serial_import
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
